add delta time into loop method in Window.java

+++ src/main/java/util/Time.java
@@ -5,6 +5,6 @@ public class Time {
 
     public static double getTime(){
 
-        return (double) (System.nanoTime() - timeStarted) * 1E-9;
+        return (System.nanoTime() - timeStarted) * 1E-9;
     }
 }
